feat: move store and router in folder "src"
