+++ src/client/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
 import App from './App.vue'
-import router from "@/client/router";
-import {store} from "@/client/store"
+import router from "@/client/src/router";
+import {store} from "@/client/src/store"
 
 createApp(App).use(router).use(store).mount('#app')
