+++ src/client/src/store/index.ts
@@ -1,10 +1,9 @@
 import { createStore } from 'vuex'
-import {testStore} from "@/client/store/modules/test.module";
+import {testStore} from "@/client/src/store/modules/test.module";
 
 export interface State {
     count: number
 }
-
 
 export const store = createStore<State>({
     state: {
